requirements.txt
aiogram==3.6.0
aiogram_dialog==2.1.0
aiohttp==3.9.0
alembic==1.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
geopy==2.4.1
pydantic==2.7.4
pydantic_settings==2.3.4
SQLAlchemy==2.0.23
transliterate==1.10.2
